requirements.txt
flask
flask_sqlalchemy
psycopg2-binary
gunicorn
PyJWT
passlib[bcrypt]
requests
python-dotenv